+++ xmake/core/base/deprecated/interpreter.lua
@@ -154,13 +154,8 @@ function deprecated_interpreter:_api_register_set_xxx_xxx(scope_kind, apiname)
     -- register api
     self:api_register_builtin(oldapi, function (value, ...) 
 
-                                        -- check
-                                        if type(value) ~= "string" then
-                                            value = type(value)
-                                        end
-
                                         -- deprecated
-                                        deprecated.add(newapi .. "(\"%s\")", oldapi .. "(\"%s\")", value)
+                                        deprecated.add(newapi .. "(\"%s\")", oldapi .. "(\"%s\")", tostring(value))
                                       
                                         -- dispatch it
                                         apifunc(value, ...)
@@ -183,13 +178,8 @@ function deprecated_interpreter:_api_register_add_xxx_xxx(scope_kind, apiname)
     -- register api
     self:api_register_builtin(oldapi, function (value, ...) 
 
-                                        -- check
-                                        if type(value) ~= "string" then
-                                            value = type(value)
-                                        end
-
                                         -- deprecated
-                                        deprecated.add(newapi .. "(\"%s\")", oldapi .. "(\"%s\")", value)
+                                        deprecated.add(newapi .. "(\"%s\")", oldapi .. "(\"%s\")", tostring(value))
                                       
                                         -- dispatch it
                                         apifunc(value, ...)