+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add debugger support on windows (vsjitdebugger, ollydbg, windbg ... )
 * Add `getenv` interface for the global scope of `xmake.lua`
 * Add plugin for generating vs200x project file 
+* Add `set_default` api for option
 
 ### Changes
 
@@ -150,6 +151,7 @@
 * 在windows上添加调试器支持（vsjitdebugger, ollydbg, windbg ... ）
 * 添加`getenv`接口到`xmake.lua`的全局作用域中
 * 添加生成vs200x工程插件
+* 为option添加`set_default`接口
 
 ### 改进